# Fix tooltip precision issues when zoomed in