--- a/crates/renderer/src/compute/ema_calculator.rs
+++ b/crates/renderer/src/compute/ema_calculator.rs
@@ -146,10 +146,7 @@ impl EmaCalculator {
         period: EmaPeriod,
         encoder: &mut wgpu::CommandEncoder,
     ) -> Result<ComputeResult, String> {
-        log::info!("[EmaCalculator] calculate_single called:");
-        log::info!("  - period: {:?} (value={})", period, period.value());
-        log::info!("  - element_count: {}", element_count);
-        log::info!("  - price_buffer: {:p}, size: {} bytes", price_buffer, price_buffer.size());
+        // Calculate EMA for period
         
         // Validate buffer size
         const MAX_BUFFER_SIZE: u64 = 256 * 1024 * 1024; // 256MB max
@@ -158,23 +155,19 @@ impl EmaCalculator {
             return Err(format!("Buffer size {} exceeds maximum allowed size {}", buffer_size, MAX_BUFFER_SIZE));
         }
         
-        log::info!("[EmaCalculator] Calculating EMA {} for {} elements", period.value(), element_count);
+        // Calculating EMA for elements
         
-        // Calculate and log alpha value for this period
-        let alpha = 2.0 / (period.value() as f32 + 1.0);
-        log::debug!("[EmaCalculator] EMA {} alpha value: {:.6}, weight of new data: {:.2}%", 
-            period.value(), alpha, alpha * 100.0);
+        // Calculate alpha value for this period
+        let _alpha = 2.0 / (period.value() as f32 + 1.0);
         
         // Log expected behavior based on data count
         if element_count < period.value() {
             log::warn!("[EmaCalculator] Only {} data points for EMA {} - will use all available data for initial average", 
                 element_count, period.value());
         } else if element_count < period.value() * 3 {
-            log::debug!("[EmaCalculator] {} data points for EMA {} - limited divergence expected (need {}+ for full divergence)", 
-                element_count, period.value(), period.value() * 3);
+            // Limited divergence expected with this data count
         } else {
-            log::debug!("[EmaCalculator] {} data points for EMA {} - sufficient data for proper EMA behavior", 
-                element_count, period.value());
+            // Sufficient data for proper EMA behavior
         }
         
         // No cache - was unsafe without data content hashing
@@ -187,12 +180,7 @@ impl EmaCalculator {
             alpha_denominator: period.value() + 1,
         };
         
-        log::info!("[EmaCalculator] EMA params:");
-        log::info!("  - period: {}", params.period);
-        log::info!("  - alpha_numerator: {}", params.alpha_numerator);
-        log::info!("  - alpha_denominator: {}", params.alpha_denominator);
-        log::info!("  - element_count: {}", params.element_count);
-        log::info!("  - alpha value: {:.6}", 2.0 / (params.alpha_denominator as f32));
+        // EMA params configured
 
         // Create a unique params buffer for this EMA calculation to avoid parameter sharing bugs
         let unique_params_buffer = self.infrastructure.device.create_buffer(&wgpu::BufferDescriptor {
@@ -236,15 +224,11 @@ impl EmaCalculator {
             });
 
         // Execute compute pass (single workgroup for sequential calculation)
-        log::info!("[EmaCalculator] Executing compute pass for EMA {}", period.value());
         self.infrastructure
             .execute_compute(encoder, &self.pipeline, &bind_group, (1, 1, 1));
 
         // Cache disabled - would need data hash to prevent stale results
         // self.cache.insert((period, element_count, data_hash), output_buffer.clone());
-
-        log::info!("[EmaCalculator] EMA {} compute pass complete, output buffer: {:p}", 
-            period.value(), &output_buffer);
         
         Ok(ComputeResult {
             output_buffer,