--- a/crates/renderer/src/drawables/mod.rs
+++ b/crates/renderer/src/drawables/mod.rs
@@ -1,5 +1,6 @@
 pub mod candlestick;
 pub mod plot;
 pub mod shared_utils;
+pub mod tooltip;
 pub mod x_axis;
 pub mod y_axis;