… Cargo.lock
… crates/renderer/src/drawables/tooltip.rs
… crates/renderer/src/shaders/tooltip.wgsl
… crates/shared-types/Cargo.toml
… crates/wasm-bridge/src/chart_engine.rs
… crates/wasm-bridge/src/lib.rs
… package.json
… web/cypress/e2e/tooltip-functionality-simplified.cy.ts
… web/cypress/e2e/tooltip-functionality.cy.ts
… web/cypress/support/commands.ts
… web/package.json
… web/src/components/chart/ChartTooltip.test.tsx
… web/src/components/chart/ChartTooltip.tsx
… web/src/components/chart/WasmCanvas.integration.test.tsx
… web/src/components/chart/WasmCanvas.tsx
… web/src/test/setup.ts